computational_litho_ai/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
starlette==0.27.0

torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
scipy==1.10.1

transformers==4.30.2
sentence-transformers==2.2.2

langchain==0.0.196
openai==0.27.6

redis==4.5.5
aioredis==2.0.1

httpx==0.24.0
requests==2.28.0

qiskit==0.41.1
pennylane==0.28.0
pennylane-qiskit==0.28.0

PyYAML==6.0
python-dotenv==1.0.0

pytest==7.3.1
pytest-asyncio==0.21.0

loguru==0.7.0
